Move the instance poll utils to provider utils
By doing so we avoid potential circular imports. Besides, those
are expected to be used by the providers only.
